+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2026-03-06
+
+### Added
+- Recursive directory support (--recursive / -r flag) for organize and dedupe commands
+- Now you can organize/dedupe files in subdirectories too!
+
 ## [1.0.0] - 2026-03-05
 
 ### Added